feat: Reduce container image size (#26)
* feat: Reduce container image size

Use python slim instead of python to reduce the base image size from 1GB
to 130MB.
Rollback from python:3.12 to python:3.11 to fix pymumble compatibility
issue.
Disable python cache building to further reduce the size of the
container image.
Use venv to manage python packages.
Use a multi-stage container build to simplify the cleanup phase.
Add an unprivileged user to run the mumbleice script.

* chore: Disable major and minor updates for python

Pymumble is not compatible with python:3.12, so renovate should not try
to bump the python version.